Add make target to run static analysis
Add a new make target 'static-analysis' to statically inspect the code.

For every object file a corresponding .plist file will be generated.
Because the operation is driven by make, a file will not be analyzed
again until it has been modified. This is useful during development when
turn-around speed is important, but it also means that to verify
everything from scratch one should execute

  $ make -B static-analysis

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>