Add xalloc, xfree wrapper functions
Add a new header, wrapper.h, to house wrapper functions around standard
stuff like malloc. Every wrapper starts with the letter x and will die()
if the wrapped function returns an error. This alleviates the need for
the caller to check the return value.

The xfree function is only for symmetry with xalloc but can be extended
for memory debugging if needed.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>