Add adb logcat binary stream reader
The reader will read logger_entry structs based on the same structure in
Android. These can be up to 5 kiB in size, so while it is efficient to
keep a buffer that size for parsing, the struct should be converted to a
more memory efficient representation before propagating to the rest of
the program.

The reader assumes logcat binary data as obtained by logcat --binary.

Issue: #1
Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>