Add logcat test data
Data obtained by running

  $ adb logcat --binary >file

on a Pixel 2 running the Android P Developer Preview.

Also add a function to allow test cases to obtain the directory in which
test data is stored. Note: this is currently hard-wired to the current
project directory structure. This implicit requirement should be relaxed
when proper command line argument parsing is introduced.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>